Time: 0 ms (100%), Space: 95.2 MB (42.82%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -43,7 +43,7 @@ public:
         return ind2ele;
     }
 
-    return ( (double) (double) ind1ele + ind2ele ) / 2.0;
+    return ( ind1ele + ind2ele ) / 2.0;
 
 
     }
